refactor: clear pr quality gates
