+++ packages/data-repositories/src/iam/seed-plan.personas.ts
@@ -1,0 +1,171 @@
+import type { PersonaSeed } from './types.js';
+import {
+  applicationReadPermissionKeys,
+  iamSeedPermissions,
+  mediaManagePermissionKeys,
+  mediaReadPermissionKeys,
+} from './seed-plan.permissions.js';
+
+export const iamSeedPersonas: readonly PersonaSeed[] = [
+  {
+    personaKey: 'system_admin',
+    roleSlug: 'system_admin',
+    roleLevel: 100,
+    displayName: 'System Administrator',
+    scopeDefault: 'instance',
+    mfaPolicy: 'required',
+    permissionKeys: iamSeedPermissions.map(([, key]) => key),
+    accountId: '50111111-1111-1111-1111-111111111111',
+    keycloakSubject: 'seed:system_admin',
+    seedEmailPlaceholder: 'seed.system_admin@sva.local',
+    seedDisplayNamePlaceholder: 'System Administrator',
+  },
+  {
+    personaKey: 'instance_registry_admin',
+    roleSlug: 'instance_registry_admin',
+    roleLevel: 95,
+    displayName: 'Instance Registry Administrator',
+    scopeDefault: 'instance',
+    mfaPolicy: 'required',
+    permissionKeys: ['instance.registry.manage', 'feature.toggle', 'integration.manage', ...applicationReadPermissionKeys],
+    accountId: '50888888-8888-8888-8888-888888888888',
+    keycloakSubject: 'seed:instance_registry_admin',
+    seedEmailPlaceholder: 'seed.instance_registry_admin@sva.local',
+    seedDisplayNamePlaceholder: 'Instance Registry Administrator',
+  },
+  {
+    personaKey: 'app_manager',
+    roleSlug: 'app_manager',
+    roleLevel: 80,
+    displayName: 'App Manager',
+    scopeDefault: 'instance',
+    mfaPolicy: 'recommended',
+    permissionKeys: [
+      'iam.user.read',
+      'iam.user.write',
+      'iam.org.read',
+      'iam.org.write',
+      'content.read',
+      'content.readHistory',
+      ...applicationReadPermissionKeys,
+      ...mediaReadPermissionKeys,
+      'feature.toggle',
+    ],
+    accountId: '50222222-2222-2222-2222-222222222222',
+    keycloakSubject: 'seed:app_manager',
+    seedEmailPlaceholder: 'seed.app_manager@sva.local',
+    seedDisplayNamePlaceholder: 'App Manager',
+  },
+  {
+    personaKey: 'feature_manager',
+    roleSlug: 'feature-manager',
+    roleLevel: 60,
+    displayName: 'Feature Manager',
+    scopeDefault: 'instance',
+    mfaPolicy: 'recommended',
+    permissionKeys: [
+      'content.read',
+      'content.readHistory',
+      ...applicationReadPermissionKeys,
+      'content.updateMetadata',
+      'content.updatePayload',
+      'content.changeStatus',
+      ...mediaManagePermissionKeys,
+      'feature.toggle',
+    ],
+    accountId: '50333333-3333-3333-3333-333333333333',
+    keycloakSubject: 'seed:feature_manager',
+    seedEmailPlaceholder: 'seed.feature-manager@sva.local',
+    seedDisplayNamePlaceholder: 'Feature Manager',
+  },
+  {
+    personaKey: 'interface_manager',
+    roleSlug: 'interface-manager',
+    roleLevel: 50,
+    displayName: 'Interface Manager',
+    scopeDefault: 'instance',
+    mfaPolicy: 'recommended',
+    permissionKeys: [
+      'iam.org.read',
+      'content.read',
+      'content.readHistory',
+      ...applicationReadPermissionKeys,
+      ...mediaReadPermissionKeys,
+      'integration.manage',
+    ],
+    accountId: '50444444-4444-4444-4444-444444444444',
+    keycloakSubject: 'seed:interface_manager',
+    seedEmailPlaceholder: 'seed.interface-manager@sva.local',
+    seedDisplayNamePlaceholder: 'Interface Manager',
+  },
+  {
+    personaKey: 'designer',
+    roleSlug: 'designer',
+    roleLevel: 40,
+    displayName: 'Designer',
+    scopeDefault: 'org',
+    mfaPolicy: 'optional',
+    permissionKeys: [
+      'content.read',
+      'content.readHistory',
+      ...applicationReadPermissionKeys,
+      'content.updateMetadata',
+      'content.updatePayload',
+      'media.read',
+      'media.update',
+    ],
+    accountId: '50555555-5555-5555-5555-555555555555',
+    keycloakSubject: 'seed:designer',
+    seedEmailPlaceholder: 'seed.designer@sva.local',
+    seedDisplayNamePlaceholder: 'Designer',
+  },
+  {
+    personaKey: 'editor',
+    roleSlug: 'editor',
+    roleLevel: 30,
+    displayName: 'Editor',
+    scopeDefault: 'org',
+    mfaPolicy: 'optional',
+    permissionKeys: [
+      'content.read',
+      'content.readHistory',
+      ...applicationReadPermissionKeys,
+      'content.create',
+      'content.updateMetadata',
+      'content.updatePayload',
+      'content.changeStatus',
+      'content.delete',
+      'media.read',
+      'media.create',
+      'media.update',
+      'media.reference.manage',
+    ],
+    accountId: '50666666-6666-6666-6666-666666666666',
+    keycloakSubject: 'seed:editor',
+    seedEmailPlaceholder: 'seed.editor@sva.local',
+    seedDisplayNamePlaceholder: 'Editor',
+  },
+  {
+    personaKey: 'moderator',
+    roleSlug: 'moderator',
+    roleLevel: 35,
+    displayName: 'Moderator',
+    scopeDefault: 'org',
+    mfaPolicy: 'optional',
+    permissionKeys: [
+      'content.read',
+      'content.readHistory',
+      ...applicationReadPermissionKeys,
+      'content.changeStatus',
+      'content.publish',
+      'content.archive',
+      'content.restore',
+      'content.manageRevisions',
+      ...mediaReadPermissionKeys,
+    ],
+    accountId: '50777777-7777-7777-7777-777777777777',
+    keycloakSubject: 'seed:moderator',
+    seedEmailPlaceholder: 'seed.moderator@sva.local',
+    seedDisplayNamePlaceholder: 'Moderator',
+  },
+];
