+++ packages/data-repositories/src/iam/seed-plan.ts
@@ -1,203 +1,6 @@
-import type { IamSeedPlan, PermissionKey, PersonaSeed } from './types.js';
-
-const permissions = [
-  ['40111111-1111-1111-1111-111111111111', 'iam.user.read', 'Read account data'],
-  ['40111111-1111-1111-1111-111111111112', 'iam.user.write', 'Modify account data'],
-  ['40111111-1111-1111-1111-111111111113', 'iam.role.read', 'Read role assignments'],
-  ['40111111-1111-1111-1111-111111111114', 'iam.role.write', 'Modify role assignments'],
-  ['40111111-1111-1111-1111-111111111115', 'iam.org.read', 'Read organization data'],
-  ['40111111-1111-1111-1111-111111111116', 'iam.org.write', 'Modify organization data'],
-  ['40111111-1111-1111-1111-111111111117', 'content.read', 'Read content'],
-  ['40111111-1111-1111-1111-111111111118', 'content.create', 'Create content'],
-  ['40111111-1111-1111-1111-111111111119', 'content.updateMetadata', 'Update content metadata'],
-  ['40111111-1111-1111-1111-111111111120', 'content.publish', 'Publish content'],
-  ['40111111-1111-1111-1111-111111111121', 'content.manageRevisions', 'Manage content revisions'],
-  ['40111111-1111-1111-1111-111111111122', 'integration.manage', 'Manage integrations'],
-  ['40111111-1111-1111-1111-111111111123', 'feature.toggle', 'Toggle feature flags'],
-  ['40111111-1111-1111-1111-111111111149', 'app.read', 'Show the app link in the sidebar'],
-  ['40111111-1111-1111-1111-111111111150', 'cockpit.read', 'Show the cockpit link in the sidebar'],
-  ['40111111-1111-1111-1111-111111111124', 'instance.registry.manage', 'Manage instance registry and provisioning'],
-  ['40111111-1111-1111-1111-111111111125', 'content.updatePayload', 'Update content payload'],
-  ['40111111-1111-1111-1111-111111111126', 'content.changeStatus', 'Change content status'],
-  ['40111111-1111-1111-1111-111111111127', 'content.archive', 'Archive content'],
-  ['40111111-1111-1111-1111-111111111128', 'content.restore', 'Restore content'],
-  ['40111111-1111-1111-1111-111111111129', 'content.readHistory', 'Read content history'],
-  ['40111111-1111-1111-1111-111111111130', 'content.delete', 'Delete content'],
-  ['40111111-1111-1111-1111-111111111143', 'media.read', 'Read media'],
-  ['40111111-1111-1111-1111-111111111144', 'media.create', 'Create media'],
-  ['40111111-1111-1111-1111-111111111145', 'media.update', 'Update media'],
-  ['40111111-1111-1111-1111-111111111146', 'media.reference.manage', 'Manage media references'],
-  ['40111111-1111-1111-1111-111111111147', 'media.delete', 'Delete media'],
-  ['40111111-1111-1111-1111-111111111148', 'media.deliver.protected', 'Deliver protected media'],
-] as const satisfies readonly [string, PermissionKey, string][];
-
-const applicationReadPermissions = ['app.read', 'cockpit.read'] as const;
-const mediaReadPermissions = ['media.read'] as const;
-const mediaManagePermissions = ['media.read', 'media.create', 'media.update', 'media.reference.manage', 'media.delete'] as const;
-
-const personas: readonly PersonaSeed[] = [
-  {
-    personaKey: 'system_admin',
-    roleSlug: 'system_admin',
-    roleLevel: 100,
-    displayName: 'System Administrator',
-    scopeDefault: 'instance',
-    mfaPolicy: 'required',
-    permissionKeys: permissions.map(([, key]) => key),
-    accountId: '50111111-1111-1111-1111-111111111111',
-    keycloakSubject: 'seed:system_admin',
-    seedEmailPlaceholder: 'seed.system_admin@sva.local',
-    seedDisplayNamePlaceholder: 'System Administrator',
-  },
-  {
-    personaKey: 'instance_registry_admin',
-    roleSlug: 'instance_registry_admin',
-    roleLevel: 95,
-    displayName: 'Instance Registry Administrator',
-    scopeDefault: 'instance',
-    mfaPolicy: 'required',
-    permissionKeys: ['instance.registry.manage', 'feature.toggle', 'integration.manage', ...applicationReadPermissions],
-    accountId: '50888888-8888-8888-8888-888888888888',
-    keycloakSubject: 'seed:instance_registry_admin',
-    seedEmailPlaceholder: 'seed.instance_registry_admin@sva.local',
-    seedDisplayNamePlaceholder: 'Instance Registry Administrator',
-  },
-  {
-    personaKey: 'app_manager',
-    roleSlug: 'app_manager',
-    roleLevel: 80,
-    displayName: 'App Manager',
-    scopeDefault: 'instance',
-    mfaPolicy: 'recommended',
-    permissionKeys: [
-      'iam.user.read',
-      'iam.user.write',
-      'iam.org.read',
-      'iam.org.write',
-      'content.read',
-      'content.readHistory',
-      ...applicationReadPermissions,
-      ...mediaReadPermissions,
-      'feature.toggle',
-    ],
-    accountId: '50222222-2222-2222-2222-222222222222',
-    keycloakSubject: 'seed:app_manager',
-    seedEmailPlaceholder: 'seed.app_manager@sva.local',
-    seedDisplayNamePlaceholder: 'App Manager',
-  },
-  {
-    personaKey: 'feature_manager',
-    roleSlug: 'feature-manager',
-    roleLevel: 60,
-    displayName: 'Feature Manager',
-    scopeDefault: 'instance',
-    mfaPolicy: 'recommended',
-    permissionKeys: [
-      'content.read',
-      'content.readHistory',
-      ...applicationReadPermissions,
-      'content.updateMetadata',
-      'content.updatePayload',
-      'content.changeStatus',
-      ...mediaManagePermissions,
-      'feature.toggle',
-    ],
-    accountId: '50333333-3333-3333-3333-333333333333',
-    keycloakSubject: 'seed:feature_manager',
-    seedEmailPlaceholder: 'seed.feature-manager@sva.local',
-    seedDisplayNamePlaceholder: 'Feature Manager',
-  },
-  {
-    personaKey: 'interface_manager',
-    roleSlug: 'interface-manager',
-    roleLevel: 50,
-    displayName: 'Interface Manager',
-    scopeDefault: 'instance',
-    mfaPolicy: 'recommended',
-    permissionKeys: [
-      'iam.org.read',
-      'content.read',
-      'content.readHistory',
-      ...applicationReadPermissions,
-      ...mediaReadPermissions,
-      'integration.manage',
-    ],
-    accountId: '50444444-4444-4444-4444-444444444444',
-    keycloakSubject: 'seed:interface_manager',
-    seedEmailPlaceholder: 'seed.interface-manager@sva.local',
-    seedDisplayNamePlaceholder: 'Interface Manager',
-  },
-  {
-    personaKey: 'designer',
-    roleSlug: 'designer',
-    roleLevel: 40,
-    displayName: 'Designer',
-    scopeDefault: 'org',
-    mfaPolicy: 'optional',
-    permissionKeys: [
-      'content.read',
-      'content.readHistory',
-      ...applicationReadPermissions,
-      'content.updateMetadata',
-      'content.updatePayload',
-      'media.read',
-      'media.update',
-    ],
-    accountId: '50555555-5555-5555-5555-555555555555',
-    keycloakSubject: 'seed:designer',
-    seedEmailPlaceholder: 'seed.designer@sva.local',
-    seedDisplayNamePlaceholder: 'Designer',
-  },
-  {
-    personaKey: 'editor',
-    roleSlug: 'editor',
-    roleLevel: 30,
-    displayName: 'Editor',
-    scopeDefault: 'org',
-    mfaPolicy: 'optional',
-    permissionKeys: [
-      'content.read',
-      'content.readHistory',
-      ...applicationReadPermissions,
-      'content.create',
-      'content.updateMetadata',
-      'content.updatePayload',
-      'content.changeStatus',
-      'content.delete',
-      'media.read',
-      'media.create',
-      'media.update',
-      'media.reference.manage',
-    ],
-    accountId: '50666666-6666-6666-6666-666666666666',
-    keycloakSubject: 'seed:editor',
-    seedEmailPlaceholder: 'seed.editor@sva.local',
-    seedDisplayNamePlaceholder: 'Editor',
-  },
-  {
-    personaKey: 'moderator',
-    roleSlug: 'moderator',
-    roleLevel: 35,
-    displayName: 'Moderator',
-    scopeDefault: 'org',
-    mfaPolicy: 'optional',
-    permissionKeys: [
-      'content.read',
-      'content.readHistory',
-      ...applicationReadPermissions,
-      'content.changeStatus',
-      'content.publish',
-      'content.archive',
-      'content.restore',
-      'content.manageRevisions',
-      ...mediaReadPermissions,
-    ],
-    accountId: '50777777-7777-7777-7777-777777777777',
-    keycloakSubject: 'seed:moderator',
-    seedEmailPlaceholder: 'seed.moderator@sva.local',
-    seedDisplayNamePlaceholder: 'Moderator',
-  },
-];
+import type { IamSeedPlan, PersonaSeed } from './types.js';
+import { iamSeedPermissions } from './seed-plan.permissions.js';
+import { iamSeedPersonas } from './seed-plan.personas.js';
 
 export const iamSeedPlan: IamSeedPlan = {
   context: {
@@ -245,8 +48,8 @@ export const iamSeedPlan: IamSeedPlan = {
       metadata: { seed: true, version: 'v2', level: 'district' },
     },
   ],
-  personas,
-  permissions: permissions.map(([id, key, description]) => ({
+  personas: iamSeedPersonas,
+  permissions: iamSeedPermissions.map(([id, key, description]) => ({
     id,
     key,
     action: key,
