+++ packages/data/scripts/test-seeds.sh
@@ -121,10 +121,10 @@ assert_count "SELECT COUNT(*) FROM iam.roles WHERE instance_id = 'de-musterhause
 assert_count "SELECT COUNT(*) FROM iam.roles WHERE instance_id = 'de-musterhausen' AND managed_by = 'studio';" "16" "studio role count"
 assert_count "SELECT COUNT(*) FROM iam.roles WHERE instance_id = 'de-musterhausen' AND role_key = 'instance_registry_admin';" "1" "instance registry admin role"
 assert_count "SELECT COUNT(*) FROM iam.roles WHERE instance_id = 'de-musterhausen' AND managed_by = 'external';" "0" "external role count"
-assert_count "SELECT COUNT(*) FROM iam.permissions WHERE instance_id = 'de-musterhausen';" "38" "permission count"
-assert_count "SELECT COUNT(*) FROM iam.permissions WHERE instance_id = 'de-musterhausen' AND action = permission_key;" "38" "permission actions"
-assert_count "SELECT COUNT(*) FROM iam.permissions WHERE instance_id = 'de-musterhausen' AND effect = 'allow';" "38" "permission effects"
-assert_count "SELECT COUNT(*) FROM iam.permissions WHERE instance_id = 'de-musterhausen' AND scope = '{}'::jsonb;" "38" "permission scopes"
+assert_count "SELECT COUNT(*) FROM iam.permissions WHERE instance_id = 'de-musterhausen';" "40" "permission count"
+assert_count "SELECT COUNT(*) FROM iam.permissions WHERE instance_id = 'de-musterhausen' AND action = permission_key;" "40" "permission actions"
+assert_count "SELECT COUNT(*) FROM iam.permissions WHERE instance_id = 'de-musterhausen' AND effect = 'allow';" "40" "permission effects"
+assert_count "SELECT COUNT(*) FROM iam.permissions WHERE instance_id = 'de-musterhausen' AND scope = '{}'::jsonb;" "40" "permission scopes"
 assert_count "SELECT COUNT(*) FROM iam.accounts WHERE keycloak_subject LIKE 'seed:%';" "8" "account count"
 assert_count "SELECT COUNT(*) FROM iam.accounts WHERE instance_id = 'de-musterhausen' AND keycloak_subject = 'seed:instance_registry_admin';" "1" "instance registry admin account"
 assert_count "SELECT COUNT(*) FROM iam.instance_memberships WHERE instance_id = 'de-musterhausen';" "8" "instance memberships"
@@ -133,9 +133,9 @@ assert_count "SELECT COUNT(*) FROM iam.account_roles WHERE instance_id = 'de-mus
 assert_count "SELECT COUNT(*) FROM iam.account_roles WHERE instance_id = 'de-musterhausen' AND account_id = '50888888-8888-8888-8888-888888888888';" "1" "instance registry admin account role"
 assert_count "SELECT COUNT(*) FROM iam.account_organizations WHERE instance_id = 'de-musterhausen';" "10" "account organizations"
 assert_count "SELECT COUNT(*) FROM iam.account_organizations WHERE instance_id = 'de-musterhausen' AND is_default_context = true;" "7" "default organization contexts"
-assert_count "SELECT COUNT(*) FROM iam.role_permissions WHERE instance_id = 'de-musterhausen';" "129" "role permissions"
-assert_count "SELECT COUNT(*) FROM iam.role_permissions WHERE instance_id = 'de-musterhausen' AND grant_origin_kind = 'manual';" "129" "manual role permissions"
-assert_count "SELECT COUNT(*) FROM iam.role_permissions rp JOIN iam.roles r ON r.id = rp.role_id AND r.instance_id = rp.instance_id WHERE rp.instance_id = 'de-musterhausen' AND r.role_key = 'instance_registry_admin';" "3" "instance registry admin role permissions"
+assert_count "SELECT COUNT(*) FROM iam.role_permissions WHERE instance_id = 'de-musterhausen';" "145" "role permissions"
+assert_count "SELECT COUNT(*) FROM iam.role_permissions WHERE instance_id = 'de-musterhausen' AND grant_origin_kind = 'manual';" "145" "manual role permissions"
+assert_count "SELECT COUNT(*) FROM iam.role_permissions rp JOIN iam.roles r ON r.id = rp.role_id AND r.instance_id = rp.instance_id WHERE rp.instance_id = 'de-musterhausen' AND r.role_key = 'instance_registry_admin';" "5" "instance registry admin role permissions"
 assert_count "SELECT COUNT(*) FROM iam.role_permissions WHERE instance_id = 'de-musterhausen' AND grant_origin_module_id IS NOT NULL;" "0" "module-owned role permissions in seeds"
 
 echo "Seed idempotency integration test passed."
