+++ packages/plugin-waste-management/src/waste-management.master-data.loaders.ts
@@ -11,6 +11,58 @@ import type { WasteMasterDataState } from './waste-management.master-data.state.
 
 type Translate = (key: string, variables?: Readonly<Record<string, string | number>>) => string;
 
+const resolveMasterDataOverviewScope = (tab: WasteManagementSearchParams['masterDataTab']) => {
+  if (tab === 'fractions') {
+    return { scope: 'fractions' } as const;
+  }
+
+  if (tab === 'locations') {
+    return { scope: 'locations' } as const;
+  }
+
+  return undefined;
+};
+
+const loadLocationOutputOverview = async (tab: WasteManagementSearchParams['masterDataTab']) => {
+  if (tab !== 'locations') {
+    return null;
+  }
+
+  try {
+    return await getWasteManagementOutputOverview();
+  } catch {
+    return null;
+  }
+};
+
+const resolveMasterDataLoadError = (translate: Translate, loadError: unknown) => {
+  const code = resolveApiErrorCode(loadError);
+  return code === 'forbidden'
+    ? translate('masterData.messages.loadForbidden')
+    : translate('masterData.messages.loadError');
+};
+
+const loadAvailableToursForLocations = async (
+  tab: WasteManagementSearchParams['masterDataTab'],
+  setAvailableTours: WasteMasterDataState['setAvailableTours'],
+  isMountedRef: React.MutableRefObject<boolean>,
+) => {
+  if (tab !== 'locations') {
+    return;
+  }
+
+  try {
+    const response = await getWasteManagementToursOverview();
+    if (isMountedRef.current) {
+      setAvailableTours(response.tours);
+    }
+  } catch {
+    if (isMountedRef.current) {
+      setAvailableTours([]);
+    }
+  }
+};
+
 export const useWasteMasterDataDataLoading = (
   state: WasteMasterDataState,
   pt: Translate,
@@ -24,30 +76,16 @@ export const useWasteMasterDataDataLoading = (
   const loadOverview = useCallback(
     async () => {
       try {
-        const overviewResponse = await getWasteManagementMasterDataOverview(
-          tab === 'fractions' ? { scope: 'fractions' } : tab === 'locations' ? { scope: 'locations' } : undefined
-        );
+        const overviewResponse = await getWasteManagementMasterDataOverview(resolveMasterDataOverviewScope(tab));
         if (!isMountedRef.current) return;
-        let outputResponse = null;
-        if (tab === 'locations') {
-          try {
-            outputResponse = await getWasteManagementOutputOverview();
-          } catch {
-            outputResponse = null;
-          }
-          if (!isMountedRef.current) return;
-        }
+        const outputResponse = await loadLocationOutputOverview(tab);
+        if (!isMountedRef.current) return;
         setOverview(overviewResponse);
         setOutputOverview(outputResponse);
         setError(null);
       } catch (loadError) {
         if (!isMountedRef.current) return;
-        const code = resolveApiErrorCode(loadError);
-        setError(
-          code === 'forbidden'
-            ? ptRef.current('masterData.messages.loadForbidden')
-            : ptRef.current('masterData.messages.loadError')
-        );
+        setError(resolveMasterDataLoadError(ptRef.current, loadError));
         setAvailableTours([]);
         setOutputOverview(null);
       } finally {
@@ -70,14 +108,7 @@ export const useWasteMasterDataDataLoading = (
       return;
     }
 
-    void (async () => {
-      try {
-        const response = await getWasteManagementToursOverview();
-        if (isMountedRef.current) setAvailableTours(response.tours);
-      } catch {
-        if (isMountedRef.current) setAvailableTours([]);
-      }
-    })();
+    void loadAvailableToursForLocations(tab, setAvailableTours, isMountedRef);
   }, [setAvailableTours, state.overview, tab]);
 
   return loadOverview;
