+++ packages/plugin-waste-management/src/waste-management.tours.content.tsx
@@ -1,5 +1,4 @@
 import { useMemo, useState } from 'react';
-import type { WasteTourRecord } from '@sva/plugin-sdk';
 import { usePluginTranslation } from '@sva/plugin-sdk';
 
 import { StatusNotice } from './waste-management.page.support.js';
@@ -11,7 +10,13 @@ import {
   useWasteToursSelectionState,
 } from './waste-management.tours.content.parts.js';
 import { WasteToursEmptyState } from './waste-management.tours.empty-state.js';
-import { formatTourRecurrence } from './waste-management.tours.presentation.js';
+import {
+  applyWasteToursFilters,
+  createLocationCountByTourId,
+  resetWasteToursFilters,
+  sortWasteTours,
+  updateWasteToursSorting,
+} from './waste-management.tours.content.helpers.js';
 import type { WasteToursSortDirection, WasteToursSortField } from './waste-management.tours.table.parts.js';
 
 export { WasteToursEmptyState };
@@ -53,48 +58,14 @@ export const WasteToursContent = ({
   const pt = usePluginTranslation('wasteManagement');
   const [sortField, setSortField] = useState<WasteToursSortField | null>(null);
   const [sortDirection, setSortDirection] = useState<WasteToursSortDirection>('asc');
-  const locationCountByTourId = useMemo(() => {
-    const counts = new Map<string, number>();
-    for (const link of masterDataOverview?.locationTourLinks ?? []) {
-      counts.set(link.tourId, (counts.get(link.tourId) ?? 0) + 1);
-    }
-    return counts;
-  }, [masterDataOverview?.locationTourLinks]);
-  const sortedTours = useMemo(() => {
-    if (!sortField) {
-      return tours;
-    }
-
-    const resolveSortValue = (tour: WasteTourRecord): string => {
-      switch (sortField) {
-        case 'name':
-          return tour.name;
-        case 'recurrence': {
-          const recurrenceValue = formatTourRecurrence(
-            pt,
-            tour.recurrence,
-            tour.customRecurrenceName,
-            tour.customRecurrenceIntervalDays
-          );
-          return recurrenceValue === '—' ? '' : recurrenceValue;
-        }
-        case 'locations':
-          return String(locationCountByTourId.get(tour.id) ?? 0).padStart(6, '0');
-        case 'status':
-          return tour.active ? 'active' : 'inactive';
-        default:
-          return '';
-      }
-    };
-
-    return [...tours].sort((left, right) => {
-      const comparison = resolveSortValue(left).localeCompare(resolveSortValue(right), 'de', {
-        numeric: true,
-        sensitivity: 'base',
-      });
-      return sortDirection === 'asc' ? comparison : comparison * -1;
-    });
-  }, [locationCountByTourId, pt, sortDirection, sortField, tours]);
+  const locationCountByTourId = useMemo(
+    () => createLocationCountByTourId(masterDataOverview?.locationTourLinks),
+    [masterDataOverview?.locationTourLinks],
+  );
+  const sortedTours = useMemo(
+    () => sortWasteTours({ tours, sortField, sortDirection, locationCountByTourId, pt }),
+    [locationCountByTourId, pt, sortDirection, sortField, tours],
+  );
   const {
     selectedTourIds,
     setSelectedTourIds,
@@ -115,6 +86,7 @@ export const WasteToursContent = ({
     draftEndDateTo,
     setDraftEndDateTo,
     hasActiveFilters,
+    syncDraftFilters,
     tourPendingDelete,
     setTourPendingDelete,
     bulkDeleteOpen,
@@ -174,27 +146,14 @@ export const WasteToursContent = ({
         hasActiveFilters={hasActiveFilters}
         onOpenCreateDialog={onOpenCreateDialog}
         onOpenFilterDialog={() => {
-          setDraftQuery(query);
-          setDraftStatus(status);
-          setDraftTourWasteFractionId(tourWasteFractionId);
-          setDraftFirstDateFrom(firstDateFrom);
-          setDraftFirstDateTo(firstDateTo);
-          setDraftEndDateFrom(endDateFrom);
-          setDraftEndDateTo(endDateTo);
+          syncDraftFilters();
           setFilterDialogOpen(true);
         }}
         onFilterDialogOpenChange={setFilterDialogOpen}
         onPageChange={onPageChange}
         onSyncPageChange={onSyncPageChange}
         onPageSizeChange={onPageSizeChange}
-        onSortChange={(field) => {
-          if (field === sortField) {
-            setSortDirection((current) => (current === 'asc' ? 'desc' : 'asc'));
-            return;
-          }
-          setSortField(field);
-          setSortDirection('asc');
-        }}
+        onSortChange={(field) => updateWasteToursSorting({ field, sortField, setSortField, setSortDirection })}
         onDraftQueryChange={setDraftQuery}
         onDraftStatusChange={setDraftStatus}
         onDraftTourWasteFractionIdChange={setDraftTourWasteFractionId}
@@ -202,32 +161,21 @@ export const WasteToursContent = ({
         onDraftFirstDateToChange={setDraftFirstDateTo}
         onDraftEndDateFromChange={setDraftEndDateFrom}
         onDraftEndDateToChange={setDraftEndDateTo}
-        onApplyFilters={() => {
-          if (onFiltersChange) {
-            onFiltersChange(
-              draftQuery,
-              draftStatus,
-              draftTourWasteFractionId,
-              draftFirstDateFrom,
-              draftFirstDateTo,
-              draftEndDateFrom,
-              draftEndDateTo,
-            );
-            setFilterDialogOpen(false);
-            return;
-          }
-          onQueryChange(draftQuery);
-          onStatusChange(draftStatus);
-          setFilterDialogOpen(false);
-        }}
-        onResetFilters={() => {
-          if (onFiltersChange) {
-            onFiltersChange('', 'all', undefined, undefined, undefined, undefined, undefined);
-            return;
-          }
-          onQueryChange('');
-          onStatusChange('all');
-        }}
+        onApplyFilters={() =>
+          applyWasteToursFilters({
+            onFiltersChange,
+            onQueryChange,
+            onStatusChange,
+            setFilterDialogOpen,
+            draftQuery,
+            draftStatus,
+            draftTourWasteFractionId,
+            draftFirstDateFrom,
+            draftFirstDateTo,
+            draftEndDateFrom,
+            draftEndDateTo,
+          })}
+        onResetFilters={() => resetWasteToursFilters({ onFiltersChange, onQueryChange, onStatusChange })}
         toggleSelectAllVisible={toggleSelectAllVisible}
         toggleSelectedTour={toggleSelectedTour}
         onOpenCalendar={onOpenCalendar}
